Fix mobile responsiveness issue for radar chart.
Previously the chart caused an overflow on mobile screen due to its size.
The solution I thought of was to replace the svg in a draggable context for
mobile users so that they are able to drag, pinch, and move the elements of
the graph within a defined area even though the graph might be larger than
the width of the available viewport.